@@ -41,15 +41,15 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^17.3.9",
+    "@angular/animations": "^18.0.3",
     "@angular/cdk": "^17.3.9",
-    "@angular/common": "^17.3.9",
-    "@angular/compiler": "^17.3.9",
-    "@angular/core": "^17.3.9",
-    "@angular/forms": "^17.3.9",
-    "@angular/localize": "^17.3.9",
-    "@angular/platform-browser": "^17.3.9",
-    "@angular/platform-browser-dynamic": "^17.3.9",
+    "@angular/common": "^18.0.3",
+    "@angular/compiler": "^18.0.3",
+    "@angular/core": "^18.0.3",
+    "@angular/forms": "^18.0.3",
+    "@angular/localize": "^18.0.3",
+    "@angular/platform-browser": "^18.0.3",
+    "@angular/platform-browser-dynamic": "^18.0.3",
     "@ng-bootstrap/ng-bootstrap": "^16.0.0",
     "@ngx-translate/core": "^15.0.0",
     "@ngx-translate/http-loader": "^8.0.0",
@@ -90,15 +90,15 @@
   },
   "devDependencies": {
     "@angular-builders/custom-webpack": "^17.0.2",
-    "@angular-devkit/build-angular": "^17.3.7",
-    "@angular-devkit/schematics": "^17.3.7",
+    "@angular-devkit/build-angular": "^18.0.4",
+    "@angular-devkit/schematics": "^18.0.4",
     "@angular-eslint/builder": "^17.4.1",
     "@angular-eslint/eslint-plugin": "^17.4.1",
     "@angular-eslint/eslint-plugin-template": "^17.4.1",
     "@angular-eslint/schematics": "^17.4.1",
     "@angular-eslint/template-parser": "^17.4.1",
-    "@angular/cli": "^17.3.7",
-    "@angular/compiler-cli": "^17.3.9",
+    "@angular/cli": "^18.0.4",
+    "@angular/compiler-cli": "^18.0.3",
     "@commitlint/cli": "^19.3.0",
     "@commitlint/config-conventional": "^19.2.2",
     "@compodoc/compodoc": "^1.1.25",
@@ -126,7 +126,7 @@
     "karma-chrome-launcher": "~3.2.0",
     "karma-jasmine": "~5.1.0",
     "karma-jasmine-html-reporter": "^2.0.0",
-    "ng-packagr": "^17.3.0",
+    "ng-packagr": "^18.0.0",
     "prettier": "^3.3.0",
     "puppeteer": "^22.10.0",
     "sass": "^1.77.4",
