@@ -1,14 +1,15 @@
 import {CUSTOM_ELEMENTS_SCHEMA, NgModule} from '@angular/core';
 import {CommonModule} from '@angular/common';
-import {FilterPipe} from 'hslayers-ng/common/pipes';
 import {FormsModule} from '@angular/forms';
+import {provideHttpClient, withInterceptorsFromDi} from '@angular/common/http';
+
+import {FilterPipe} from 'hslayers-ng/common/pipes';
 import {HsLegendComponent} from './legend.component';
 import {HsLegendLayerComponent} from './legend-layer/legend-layer.component';
 import {HsLegendLayerStaticComponent} from './legend-layer-static/legend-layer-static.component';
 import {HsLegendLayerVectorComponent} from './legend-layer-vector/legend-layer-vector.component';
 import {HsPanelHeaderComponent} from 'hslayers-ng/common/panels';
 import {HsPanelHelpersModule} from 'hslayers-ng/common/panels';
-import {HttpClientModule} from '@angular/common/http';
 import {TranslateCustomPipe} from 'hslayers-ng/services/language';
 @NgModule({
   schemas: [CUSTOM_ELEMENTS_SCHEMA],
@@ -21,12 +22,12 @@ import {TranslateCustomPipe} from 'hslayers-ng/services/language';
   imports: [
     CommonModule,
     FormsModule,
-    HttpClientModule,
     HsPanelHelpersModule,
     FilterPipe,
     TranslateCustomPipe,
     HsPanelHeaderComponent,
   ],
   exports: [HsLegendComponent, HsLegendLayerComponent],
+  providers: [provideHttpClient(withInterceptorsFromDi())],
 })
 export class HsLegendModule {}
